+++ modules/P02_firmware_bin_file_check.sh
@@ -25,35 +25,40 @@ P02_firmware_bin_file_check() {
   export VMDK_DETECTED=0
   export DLINK_ENC_DETECTED=0
 
-  FILE_BIN_OUT=$(file "$FIRMWARE_PATH")
-  DLINK_ENC_CHECK=$(hexdump -C "$FIRMWARE_PATH"| head -1)
+  if [[ -f "$FIRMWARE_PATH" ]]; then
+    FILE_BIN_OUT=$(file "$FIRMWARE_PATH")
+    DLINK_ENC_CHECK=$(hexdump -C "$FIRMWARE_PATH"| head -1)
 
-  if [[ "$FILE_BIN_OUT" == *"VMware4 disk image"* ]]; then
-    export VMDK_DETECTED=1
-  fi
-  if [[ "$DLINK_ENC_CHECK" == *"SHRS"* ]]; then
-    export DLINK_ENC_DETECTED=1
+    if [[ "$FILE_BIN_OUT" == *"VMware4 disk image"* ]]; then
+      export VMDK_DETECTED=1
+    fi
+    if [[ "$DLINK_ENC_CHECK" == *"SHRS"* ]]; then
+      export DLINK_ENC_DETECTED=1
+    fi
+
+    # entropy checking on binary file
+    ENTROPY=$(ent "$FIRMWARE_PATH" | grep Entropy)
   fi
 
   local FILE_LS_OUT
   FILE_LS_OUT=$(ls -lh "$FIRMWARE_PATH")
-
-  # entropy checking on binary file
-  ENTROPY=$(ent "$FIRMWARE_PATH" | grep Entropy)
   
   print_output "[*] Details of the binary file:"
   print_output "$(indent "$FILE_LS_OUT")"
   print_output ""
-  print_output "$(indent "$FILE_BIN_OUT")"
-  print_output ""
-  print_output "$(indent "$ENTROPY")"
-  print_output ""
-  if [[ -x "$EXT_DIR"/pixde ]]; then
-    print_output "[*] Visualized firmware file (first 2000 bytes):"
-    "$EXT_DIR"/pixde -r-0x2000 "$FIRMWARE_PATH" | tee -a "$LOG_DIR"/p02_pixd.txt
+
+  if [[ -f "$FIRMWARE_PATH" ]]; then
+    print_output "$(indent "$FILE_BIN_OUT")"
     print_output ""
-    python3 "$EXT_DIR"/pixd_png.py -i "$LOG_DIR"/p02_pixd.txt -o "$LOG_DIR"/pixd.png -p 10 > /dev/null
-    write_link "$LOG_DIR"/pixd.png
+    print_output "$(indent "$ENTROPY")"
+    print_output ""
+    if [[ -x "$EXT_DIR"/pixde ]]; then
+      print_output "[*] Visualized firmware file (first 2000 bytes):"
+      "$EXT_DIR"/pixde -r-0x2000 "$FIRMWARE_PATH" | tee -a "$LOG_DIR"/p02_pixd.txt
+      print_output ""
+      python3 "$EXT_DIR"/pixd_png.py -i "$LOG_DIR"/p02_pixd.txt -o "$LOG_DIR"/pixd.png -p 10 > /dev/null
+      write_link "$LOG_DIR"/pixd.png
+    fi
   fi
 
   module_end_log "${FUNCNAME[0]}" 1
